+++ numba_utils/algorithms/__init__.py
@@ -8,8 +8,10 @@ from numba_utils.algorithms._select import (
 from numba_utils.algorithms._sort import (
     counting_sort,
     insertion_sort,
+    lexsort,
     partial_sort,
     radix_sort,
+    stable_argsort,
 )
 from numba_utils.algorithms._topk import argmax2, topk
 
@@ -18,9 +20,11 @@ __all__ = [
     "counting_sort",
     "fast_argpartition",
     "insertion_sort",
+    "lexsort",
     "nth_element",
     "partial_sort",
     "quickselect",
     "radix_sort",
+    "stable_argsort",
     "topk",
 ]